src/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-crontab==0.7.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djoser==2.0.2
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.3
psycopg2-binary==2.8.3
pypugjs==5.8.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
